Fix FMOD_SYS_FMOD_DIRECTORY envvar not triggering rebuild
Previously, setting or changing the FMOD_SYS_FMOD_DIRECTORY envvar after
doing an initial (successful) build of this crate would not trigger a
rebuild - so if you changed the envvar to point to a different version
of FMOD, then you'd actually still be using the old version of FMOD
without knowing it.

This fixes that by instructing Cargo to rebuild this crate if the envvar
changes.