Make release() fns take &self instead of self as arg
Since the various types that can be released are Copy (i.e. we're
treating them as basically being pointers), it doesn't really make sense
to have release() take ownership of self.

The alternative approach would be to remove the Copy impl from basically
all the FMOD types so that we treat them as resources rather than
pointers - and then maybe impl Drop for them? (at least for all except
system).

Either way, good to pick an approach and stick to it.